+++ CopilotKit/packages/react-core/src/hooks/use-make-copilot-actionable.ts
@@ -1,5 +1,3 @@
-"use client";
-
 import { useRef, useContext, useEffect, useMemo } from "react";
 import { CopilotContext } from "../context/copilot-context";
 import { AnnotatedFunction } from "@copilotkit/shared";