+++ CopilotKit/packages/react-textarea/src/components/ui/command.tsx
@@ -1,10 +1,6 @@
-"use client";
-
 import * as React from "react";
 import { DialogProps } from "@radix-ui/react-dialog";
 import { Command as CommandPrimitive } from "cmdk";
-import { Search } from "lucide-react";
-
 import { cn } from "../../lib/utils";
 import { Dialog, DialogContent } from "./dialog";