+++ CopilotKit/packages/react-ui/src/components/chat/CodeBlock.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { FC, memo } from "react";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
 import { vscDarkPlus as highlightStyle } from "react-syntax-highlighter/dist/esm/styles/prism";