+++ CopilotKit/examples/node/CHANGELOG.md
@@ -1,5 +1,43 @@
 # node
 
+## 1.19.0
+
+### Minor Changes
+
+- 899aa6e: Backend improvements for running on GCP
+- Improve streamHttpServerResponse for express and firebase apps
+
+### Patch Changes
+
+- Updated dependencies [899aa6e]
+- Updated dependencies
+  - @copilotkit/backend@0.7.0
+  - @copilotkit/shared@0.7.0
+
+## 1.19.0-mme-firebase-fixes.0
+
+### Minor Changes
+
+- Backend improvements for running on GCP
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/backend@0.7.0-mme-firebase-fixes.0
+  - @copilotkit/shared@0.7.0-mme-firebase-fixes.0
+
+## 1.18.0
+
+### Minor Changes
+
+- Improve Next.js support and action rendering
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/backend@0.6.0
+  - @copilotkit/shared@0.6.0
+
 ## 1.17.0
 
 ### Minor Changes