+++ CopilotKit/packages/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,24 @@
 # eslint-config-custom
 
+## 0.8.0
+
+### Minor Changes
+
+- 899aa6e: Backend improvements for running on GCP
+- Improve streamHttpServerResponse for express and firebase apps
+
+## 0.8.0-mme-firebase-fixes.0
+
+### Minor Changes
+
+- Backend improvements for running on GCP
+
+## 0.7.0
+
+### Minor Changes
+
+- Improve Next.js support and action rendering
+
 ## 0.6.0
 
 ### Minor Changes