+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,40 @@
 # ui
 
+## 0.23.0
+
+### Minor Changes
+
+- 899aa6e: Backend improvements for running on GCP
+- Improve streamHttpServerResponse for express and firebase apps
+
+### Patch Changes
+
+- Updated dependencies [899aa6e]
+- Updated dependencies
+  - @copilotkit/shared@0.7.0
+
+## 0.23.0-mme-firebase-fixes.0
+
+### Minor Changes
+
+- Backend improvements for running on GCP
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.7.0-mme-firebase-fixes.0
+
+## 0.22.0
+
+### Minor Changes
+
+- Improve Next.js support and action rendering
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@0.6.0
+
 ## 0.21.0
 
 ### Minor Changes